pyproject.toml
[project]
name = "vitalagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chromadb>=1.5.5",
    "langchain>=1.2.15",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.2.1",
    "langchain-openai>=1.1.12",
    "langgraph>=1.1.6",
    "onnxruntime<1.24",
    "pandas>=2.3.3",
    "pydantic>=2.12.5",
    "pypdf>=6.9.2",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.3.0",
    "sqlalchemy>=2.0.49",
    "streamlit>=1.56.0",
    "structlog>=25.5.0",
    "torchvision>=0.26.0",
]
